issue #27 [JavaScript Improvements] - Organize code of module navBarManager.js
