+++ FRONTEND/src/main.js
@@ -11,22 +11,24 @@ import "jquery-ui/themes/base/slider.css";
 import "vis-network/dist/dist/vis-network.min.css";
 import "./styles/style.css";
 
-//JSON
+// JSON
 import sampleConfig from "./config.json";
 import OccurData from "../../DB/RelationshipSliderData.json";
 import YearData from "../../DB/YearSliderData.json";
 import ToolTopicData from "../../DB/ToolTopicAutocomplete.json";
 import communityData from "../../DB/CommunityData.json";
 
-//Images
+// Images
 import ToolImage from "./images/tool_centered_sm.png";
 import PaperImage from "./images/paper_centered_sm.png";
 import DatabaseImage from "./images/database_centered_sm.png";
 import TopicImage from "./images/topic_centered_sm.png";
 import CloseButton from "./images/xmark-solid.svg";
-import MenuButton from "./images/bars-solid.svg";
 import LoadingIcon from "./images/spinner-solid.svg";
 import logoInSoLiTo from "./images/logo_InSoLiTo.png";
+
+// Modules
+import { actionSidebar, Barchart, sliderRangeFunction, removeLegend, addLegend, initAutocomplete } from "./modules.js/navBar";
 
 // Neovis.js options
 var Vis;
@@ -87,36 +89,6 @@ function drawVis() {
   Vis = new vis.Network(container, data, options);
 }
 
-// Open and close sidebar
-function actionSidebar() {
-  if (document.getElementById("MenuImage")) {
-    document
-      .getElementById("MenuImage")
-      .parentElement.removeChild(document.getElementById("MenuImage"));
-  }
-  var main = document.getElementById("main");
-  var button = document.getElementById("openbtn");
-  var buttonImage = document.createElement("img");
-  buttonImage.id = "MenuImage";
-  buttonImage.alt = "";
-  if (main.style.marginRight === "0px" || !main.style.marginRight) {
-    document.getElementById("mySidebar").style.width = "300px";
-    document.getElementById("mySidebar").style.paddingLeft = "10px";
-    document.getElementById("main").style.marginRight = "300px";
-    //   button.style.background = 'url('+ CloseButton+ ')';
-    buttonImage.src = CloseButton;
-    document.getElementById("visualization").style.width = "calc(100% - 300px)";
-  } else {
-    document.getElementById("mySidebar").style.width = "0";
-    document.getElementById("mySidebar").style.paddingLeft = "0";
-    document.getElementById("main").style.marginRight = "0";
-    //   button.innerHTML = '☰';
-    buttonImage.src = MenuButton;
-    document.getElementById("visualization").style.width = "100%";
-  }
-  button.appendChild(buttonImage);
-}
-
 var navButton = document.getElementById("openbtn");
 navButton.addEventListener("click", () => {
   actionSidebar();
@@ -138,98 +110,14 @@ function createHomePage() {
   homePage.insertBefore(divHomePage, homePage.firstChild);
 }
 
-window.onload = createHomePage();
-
-window.onload = removeLoadingPage();
-
-window.onload = drawVis();
-
-window.onload = actionSidebar();
-
-// Barchart functions
-function drawLine(ctx, startX, startY, endX, endY, color) {
-  ctx.save();
-  ctx.strokeStyle = color;
-  ctx.beginPath();
-  ctx.moveTo(startX, startY);
-  ctx.lineTo(endX, endY);
-  ctx.stroke();
-  ctx.restore();
-}
-
-function drawBar(
-  ctx,
-  upperLeftCornerX,
-  upperLeftCornerY,
-  width,
-  height,
-  color
-) {
-  ctx.save();
-  ctx.fillStyle = color;
-  ctx.fillRect(upperLeftCornerX, upperLeftCornerY, width, height);
-  ctx.restore();
-}
-
-var Barchart = function (options) {
-  this.options = options;
-  this.canvas = options.canvas;
-  this.ctx = this.canvas.getContext("2d");
-  this.colors = options.colors;
-
-  this.draw = function () {
-    var maxValue = 0;
-    for (var categ in this.options.data) {
-      maxValue = Math.max(maxValue, this.options.data[categ]);
-    }
-    var canvasActualHeight = this.canvas.height - this.options.padding * 2;
-    var canvasActualWidth = this.canvas.width - this.options.padding * 2;
-
-    //drawing the grid lines
-    var gridValue = 0;
-    while (gridValue <= maxValue) {
-      var gridY =
-        canvasActualHeight * (1 - gridValue / maxValue) + this.options.padding;
-      drawLine(
-        this.ctx,
-        0,
-        gridY,
-        this.canvas.width,
-        gridY,
-        this.options.gridColor
-      );
-
-      //writing grid markers
-      this.ctx.save();
-      this.ctx.fillStyle = this.options.gridColor;
-      //             this.ctx.font = "bold 10px Arial";
-      //             this.ctx.fillText(gridValue, 10,gridY - 2);
-      this.ctx.restore();
-
-      gridValue += this.options.gridScale;
-    }
-
-    //drawing the bars
-    var barIndex = 0;
-    var numberOfBars = Object.keys(this.options.data).length;
-    var barSize = canvasActualWidth / numberOfBars;
-
-    for (categ in this.options.data) {
-      var val = this.options.data[categ];
-      var barHeight = Math.round((canvasActualHeight * val) / maxValue);
-      drawBar(
-        this.ctx,
-        this.options.padding + barIndex * barSize,
-        this.canvas.height - barHeight - this.options.padding,
-        barSize,
-        barHeight,
-        this.colors[barIndex % this.colors.length]
-      );
-
-      barIndex++;
-    }
-  };
-};
+$(function () {
+  createHomePage();
+  removeLoadingPage();
+  drawVis();
+  actionSidebar();
+  sliderRangeFunction();
+  initAutocomplete(ToolTopicData, addNodes, ToolImage, DatabaseImage, TopicImage);
+});
 
 var YearCanvas = document.getElementById("YearCanvas");
 // Barchart options
@@ -252,75 +140,6 @@ var OccurBarchart = new Barchart({
 });
 // Initialize Barchart
 OccurBarchart.draw();
-
-// Function to scale the horitzontal values of the range slider
-function logslider(position) {
-  // position will be between 0 and 100
-  var minp = 0;
-  var maxp = 100;
-
-  // The result should be between 100 an 10000000
-  var minv = Math.log(parseInt(Object.keys(OccurData)[0]));
-  var maxv = Math.log(
-    parseInt(Object.keys(OccurData)[Object.keys(OccurData).length - 1])
-  );
-
-  // calculate adjustment factor
-  var scale = (maxv - minv) / (maxp - minp);
-
-  return Math.trunc(Math.exp(minv + scale * (position - minp)));
-}
-
-// Slider range function
-$(function () {
-  $("#year-slider-range").slider({
-    range: true,
-    min: parseInt(Object.keys(YearData)[0]),
-    max: parseInt(Object.keys(YearData)[Object.keys(YearData).length - 1]),
-    values: [
-      parseInt(Object.keys(YearData)[0]),
-      parseInt(Object.keys(YearData)[Object.keys(YearData).length - 1]),
-    ],
-    slide: function (event, ui) {
-      $("#yearAmount").val(ui.values[0] + " - " + ui.values[1]);
-    },
-    // When slider range changes, update the nodes
-    change: function () {
-      updateNodes();
-    },
-    create: function () {
-      $("#yearAmount").val(
-        $("#year-slider-range").slider("values", 0) +
-          " - " +
-          $("#year-slider-range").slider("values", 1)
-      );
-    },
-  });
-
-  // Slider range function
-  $("#occur-slider-range").slider({
-    range: true,
-    min: 0,
-    max: 100,
-    values: [20, 100],
-    slide: function (event, ui) {
-      $("#occurAmount").val(
-        logslider(ui.values[0]) + " - " + logslider(ui.values[1])
-      );
-    },
-    // When slider range changes, update the nodes
-    change: function () {
-      updateNodes();
-    },
-    create: function () {
-      $("#occurAmount").val(
-        logslider($("#occur-slider-range").slider("values", 0)) +
-          " - " +
-          logslider($("#occur-slider-range").slider("values", 1))
-      );
-    },
-  });
-});
 
 // Function to update InSoLiTo everytime the range slider changes
 function updateNodes() {
@@ -348,85 +167,8 @@ function updateNodes() {
     addNodes(nameNode, listNode[0], listNode[1]);
   }
 }
-
-// Autcomplete Function for the Search box
-$(function () {
-  $("#tooltopic_autocomplete")
-    .autocomplete({
-      source: function (request, response) {
-        // Escape regex
-        var term = $.ui.autocomplete.escapeRegex(request.term);
-        // Search results that start with the search term
-        var matcher1 = new RegExp("^" + term, "i");
-        // Search results that start differently
-        var matcher2 = new RegExp("^.+" + term, "i");
-
-        function subarray(matcher) {
-          return $.grep(ToolTopicData, function (item) {
-            return matcher.test(item.value);
-          });
-        }
-        response($.merge(subarray(matcher1), subarray(matcher2)));
-      },
-      minLength: 1,
-      select: function (event, ui) {
-        // Select Name and Id of tool
-        var name = ui.item.value;
-        var idNode = ui.item.idNodes;
-        var labelNode = ui.item.labelnode;
-        if (Array.isArray(labelNode)) {
-          var labelNode = labelNode[0];
-        }
-        //Add Nodes from the autocomplete
-        addNodes(name, idNode, labelNode);
-        $(this).val("");
-        return false;
-      },
-      open: function () {
-        $(".ui-autocomplete").css("z-index", 1000);
-      },
-    }) // Output of the textbox
-    .autocomplete("instance")._renderItem = function (ul, item) {
-    if (item.labelnode[0] === "Tool") {
-      return $(
-        '<li><div class="boxAutocomplete"><img src="' +
-          ToolImage +
-          '"><div><div class="TextAutocomplete">' +
-          item.value +
-          '</div><div class="typeSoft">' +
-          item.type.join("/") +
-          "</div></div></div></li>"
-      ).appendTo(ul);
-    } else if (item.labelnode[0] === "Database") {
-      return $(
-        '<li><div class="boxAutocomplete"><img src="' +
-          DatabaseImage +
-          '"><div><div class="TextAutocomplete">' +
-          item.value +
-          '</div><div class="typeSoft">' +
-          item.type.join("/") +
-          "</div></div></div></li>"
-      ).appendTo(ul);
-    } else {
-      return $(
-        '<li><div class="boxAutocomplete"><img src="' +
-          TopicImage +
-          '"><div class="TextAutocomplete">' +
-          item.value +
-          "</div></div></li>"
-      ).appendTo(ul);
-    }
-  };
-});
-
-// Empty the legend
-function removeLegend() {
-  const list = document.querySelector("#legend div");
-  list.innerHTML = "";
-}
-
 // Function that retrieves the id and size of the communities from the graph
-function returnClusters() {
+export function returnClusters() {
   var net = Vis.body;
   var allNodes = net.nodeIndices;
 
@@ -469,56 +211,6 @@ function returnClusters() {
     }
   });
   return dictClusters;
-}
-
-// Insert the legend in the HTML
-function addLegend() {
-  var optionRadio = document.querySelector(
-    'input[name="cluster_mode"]:checked'
-  );
-  const list = document.querySelector("#legend div");
-  // If normal colors
-  if (optionRadio.value === "Normal") {
-    // Insert the different type of nodes in the legend (Publication, Tool, Dataset)
-
-    list.innerHTML =
-      '<div id="legendnormal"><span id="ExpandedNode" style="background-color:#fbba7e;"></span><span> Expanded node </span></div>';
-    list.innerHTML +=
-      '<div id="legendnormal"><img style="background-color: #add8e6;" src=' +
-      ToolImage +
-      " ><span> Tools </span></div>";
-    list.innerHTML +=
-      '<div id="legendnormal"><img style="background-color: #FB7E81;" src=' +
-      PaperImage +
-      "><span> Articles </span></div>";
-    list.innerHTML +=
-      '<div id="legendnormal"><img style="background-color: #b2e6ad;" src=' +
-      DatabaseImage +
-      "><span> Databases </span></div>";
-  }
-  // If Cluster mode, you take the colors from each community
-  // If there are less than 10 nodes, don't write the community in the legend
-  else {
-    list.innerHTML = "";
-    // Retrieve community ids and their size
-    var dictClusters = returnClusters();
-
-    var listCom = [];
-    for (const [, cvalue] of Object.entries(dictClusters)) {
-      listCom.push(Object.values(cvalue));
-    }
-    var sortedArray = listCom.sort(function (a, b) {
-      return b[0] - a[0];
-    });
-    sortedArray.forEach((com) => {
-      list.innerHTML +=
-        '<div><div id="circle" style="background-color:' +
-        com[5] +
-        ';"></div><span>' +
-        com[1] +
-        "</span></div>";
-    });
-  }
 }
 
 // Function that store the community color of the nodes
